Updated Pytorch model and feature extraction
The Pytorch model's initialization has been updated to include a check for MPS availability. The input size in the constructor has been increased from 3 to 5. Additionally, the device selection logic now includes an option for 'mps' if both MPS and CUDA are available.

In terms of feature extraction, two new time features - 'month' and 'weekday' - have been added to the dataset. Consequently, these changes also affect the list of feature columns used in training the model.