Added metadata to ML models
In this update, we've introduced metadata to our machine learning models. This includes the author's name and additional information such as the model type. We've also refactored how models are initialized for a cleaner codebase. Now, when writing the model to a URL, we include this new metadata. These changes affect all classifier types: BoostedTree, DecisionTree, LogisticRegression, and RandomForest.