Updated API and subproject references
In this update, we've made a few key changes to improve functionality. The API now runs on port 4201 instead of the default, providing more flexibility for deployment. We've also updated binary files in the iOS project, which should enhance user interface responsiveness. Lastly, we've updated the commit reference for the DisaggregatePowerWebApp subproject to ensure it's pointing to the latest version.